update hw2 - casting to "unaigned long long"

+++ hw2/podem/main.cc
@@ -30,7 +30,7 @@ int realRandom(int size) {
     // return int((rand()/(RAND_MAX + 1.0))*size);
     const unsigned long long A = 1664525,
                    C = 1013904223,
-                   M = (unsigned)(pow(2, 32));
+                   M = (unsigned long long)(pow(2, 32));
     static unsigned long long num = rand() % M;
     num = (A * num + C) % M;
     return  int(size*num/double(M+1));
